feat: implement display correct players data on home match card
